+++ rmw_fastrtps_common/include/rmw_fastrtps_common/ServiceTypeSupport.h
@@ -12,19 +12,18 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef RMW_FASTRTPS_CPP__SERVICETYPESUPPORT_H_
-#define RMW_FASTRTPS_CPP__SERVICETYPESUPPORT_H_
+#ifndef RMW_FASTRTPS_COMMON__SERVICETYPESUPPORT_H_
+#define RMW_FASTRTPS_COMMON__SERVICETYPESUPPORT_H_
 
 #include <fastcdr/FastBuffer.h>
 #include <fastcdr/Cdr.h>
 #include <cassert>
 
 #include "TypeSupport.h"
-#include "rosidl_typesupport_introspection_cpp/field_types.hpp"
 
 struct CustomServiceInfo;
 
-namespace rmw_fastrtps_cpp
+namespace rmw_fastrtps_common
 {
 
 template<typename MembersType>
@@ -48,8 +47,8 @@ public:
   explicit ResponseTypeSupport(const ServiceMembersType * members);
 };
 
-}  // namespace rmw_fastrtps_cpp
+}  // namespace rmw_fastrtps_common
 
 #include "ServiceTypeSupport_impl.h"
 
-#endif  // RMW_FASTRTPS_CPP__SERVICETYPESUPPORT_H_
+#endif  // RMW_FASTRTPS_COMMON__SERVICETYPESUPPORT_H_
