+++ rmw_fastrtps_common/include/rmw_fastrtps_common/ServiceTypeSupport_impl.h
@@ -12,18 +12,17 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef RMW_FASTRTPS_CPP__SERVICETYPESUPPORT_IMPL_H_
-#define RMW_FASTRTPS_CPP__SERVICETYPESUPPORT_IMPL_H_
+#ifndef RMW_FASTRTPS_COMMON__SERVICETYPESUPPORT_IMPL_H_
+#define RMW_FASTRTPS_COMMON__SERVICETYPESUPPORT_IMPL_H_
 
 #include <fastcdr/FastBuffer.h>
 #include <fastcdr/Cdr.h>
 #include <cassert>
 #include <string>
 
-#include "rmw_fastrtps_cpp/ServiceTypeSupport.h"
-#include "rosidl_typesupport_introspection_cpp/field_types.hpp"
+#include "rmw_fastrtps_common/ServiceTypeSupport.h"
 
-namespace rmw_fastrtps_cpp
+namespace rmw_fastrtps_common
 {
 
 template<typename MembersType>
@@ -71,6 +70,6 @@ ResponseTypeSupport<ServiceMembersType, MessageMembersType>::ResponseTypeSupport
   }
 }
 
-}  // namespace rmw_fastrtps_cpp
+}  // namespace rmw_fastrtps_common
 
-#endif  // RMW_FASTRTPS_CPP__SERVICETYPESUPPORT_IMPL_H_
+#endif  // RMW_FASTRTPS_COMMON__SERVICETYPESUPPORT_IMPL_H_
