+++ rmw_fastrtps_common/include/rmw_fastrtps_common/macros.hpp
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef RMW_FASTRTPS_CPP__MACROS_HPP_
-#define RMW_FASTRTPS_CPP__MACROS_HPP_
+#ifndef RMW_FASTRTPS_COMMON__MACROS_HPP_
+#define RMW_FASTRTPS_COMMON__MACROS_HPP_
 
 #include <limits>
 #include <string>
@@ -33,4 +33,4 @@
     } \
   };
 
-#endif  // RMW_FASTRTPS_CPP__MACROS_HPP_
+#endif  // RMW_FASTRTPS_COMMON__MACROS_HPP_
